+++ project-folder/src/brain-model/imgs/brain_base.jpg
