+++ project-folder/src/brain-model/imgs/brain_cerebellum.jpg
