+++ project-folder/src/brain-model/imgs/brain_cerebrum_l_alt.jpg
