+++ project-folder/src/brain-model/imgs/brain_cerebrum_r_alt.jpg
