+++ project-folder/src/brain-model/imgs/brain_internal_alt.jpg
